Do not compare Ap[plications
